Update social media badges in README.md
Updated and added social media badges to the README file of the package `api-sdk-kit`. Modifications include removal of the flat-square style from packagist badges and addition of star count, discord, and twitter follow badges. This change aims to provide easier access to our social platforms and to get a sense of popularity from the star count.